tool/AutoDroid-V2/requirements.txt
python-dotenv==1.1.0
PyYAML==6.0.1
bs4==0.0.1
openai==1.47.0
tiktoken==0.5.2
requests==2.32.3
pandas==2.2.3
lxml==5.1.0
pillow==9.2.0
numpy==1.26.2
torch==2.2.2
sentence-transformers==2.2.2
matplotlib==3.7.2
networkx==3.2.1
pyvis==0.3.2
setuptools==78.0.2
androguard==4.1.3